chore: fix filenames for license rules
