+++ packages/core/src/rules/oas2/index.ts
@@ -4,8 +4,8 @@ import { NoInvalidSchemaExamples } from '../common/no-invalid-schema-examples';
 import { NoInvalidParameterExamples } from '../common/no-invalid-parameter-examples';
 import { InfoDescription } from '../common/info-description';
 import { InfoContact } from '../common/info-contact';
-import { InfoLicense } from '../common/info-license-url';
-import { InfoLicenseUrl } from '../common/license-url';
+import { InfoLicense } from '../common/info-license';
+import { InfoLicenseUrl } from '../common/info-license-url';
 import { BooleanParameterPrefixes } from './boolean-parameter-prefixes';
 import { TagDescription } from '../common/tag-description';
 import { TagsAlphabetical } from '../common/tags-alphabetical';
